Framework/requirements/common.txt
requests==2.12.5
psutil==5.3.1
pypiwin32==219
lxml==3.7.3
selenium==3.9.0
robotframework==3.0.2
robotframework-debuglibrary==1.1.2
robotframework-jsonlibrary==0.2
pypiwin32==219
paramiko==2.4.1
gevent==1.2.2
SendKeys==0.3
pymssql==2.1.3
PyMySQL==0.9.0
numpy
matplotlib
scipy
soundfile
pythonnet